App/requirements.txt
# CROC FPGA Monitor - Python Dependencies
# ========================================
# Install with: pip install -r requirements.txt

# Qt Framework for GUI
PySide6>=6.5.0

# Serial communication
pyserial>=3.5

# Optional: For enhanced serial port detection on some systems
# pyserial-asyncio>=0.6
